requirements.txt
streamlit
mediapipe
numpy
opencv-python
python-dotenv
